+++ sandbox/modules/identity/src/main/java/org/opensearch/identity/authz/PermissionFactory.java
@@ -57,6 +57,7 @@ public class PermissionFactory {
     public OpenSearchPermission createPermission(String permissionString) {
 
         OpenSearchPermission newPermission = new OpenSearchPermission(permissionString);
+
         permissionIsValidFormat(newPermission);
         return newPermission;
     }