+++ sandbox/modules/identity/src/main/java/org/opensearch/identity/rest/action/CheckPermissionAction.java
@@ -1,0 +1,29 @@
+/*
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * The OpenSearch Contributors require contributions made to
+ * this file be licensed under the Apache-2.0 license or a
+ * compatible open source license.
+ */
+
+package org.opensearch.identity.rest.action;
+
+import org.opensearch.action.ActionType;
+import org.opensearch.identity.rest.response.CheckPermissionResponse;
+
+/**
+ * This class defines the CheckPermissionAction ActionType which corresponds to an action that checks if a subject has a specific permission while using Identity.
+ */
+public class CheckPermissionAction extends ActionType<CheckPermissionResponse> {
+
+    public static final CheckPermissionAction INSTANCE = new CheckPermissionAction();
+
+    // TODO : revisit this action type
+
+    // The action name
+    public static final String NAME = "cluster:permission/check";
+
+    CheckPermissionAction() {
+        super(NAME, CheckPermissionResponse::new);
+    }
+}